+++ .storybook/storybook.requires.js
@@ -69,7 +69,6 @@ const getStories = () => {
     "./modules/connection/components/button/CodeEntryButton.stories.tsx": require("../modules/connection/components/button/CodeEntryButton.stories.tsx"),
     "./modules/connection/components/description/CodeDescription.stories.tsx": require("../modules/connection/components/description/CodeDescription.stories.tsx"),
     "./modules/connection/components/entranceInput/EntranceInput.stories.tsx": require("../modules/connection/components/entranceInput/EntranceInput.stories.tsx"),
-    "./modules/post/components/button/ConfirmCancelButton.stories.tsx": require("../modules/post/components/button/ConfirmCancelButton.stories.tsx"),
     "./modules/post/components/card/Card.stories.tsx": require("../modules/post/components/card/Card.stories.tsx"),
     "./modules/post/components/card/info/display/Display.stories.tsx": require("../modules/post/components/card/info/display/Display.stories.tsx"),
     "./modules/post/components/card/info/heading/BottomHeading.stories.tsx": require("../modules/post/components/card/info/heading/BottomHeading.stories.tsx"),
@@ -110,6 +109,7 @@ const getStories = () => {
     "./modules/post/screens/MeetingPostViewer.stories.tsx": require("../modules/post/screens/MeetingPostViewer.stories.tsx"),
     "./modules/shared/components/avatar/Avatar.stories.tsx": require("../modules/shared/components/avatar/Avatar.stories.tsx"),
     "./modules/shared/components/button/Button.stories.tsx": require("../modules/shared/components/button/Button.stories.tsx"),
+    "./modules/shared/components/button/ConfirmCancelButton.stories.tsx": require("../modules/shared/components/button/ConfirmCancelButton.stories.tsx"),
     "./modules/shared/components/button/IconButton.stories.tsx": require("../modules/shared/components/button/IconButton.stories.tsx"),
     "./modules/shared/components/calendar/Calendar.stories.tsx": require("../modules/shared/components/calendar/Calendar.stories.tsx"),
     "./modules/shared/components/description/Description.stories.tsx": require("../modules/shared/components/description/Description.stories.tsx"),
