+++ modules/account/components/accountManagement/accountDeletionModal/type.ts
@@ -1,0 +1,6 @@
+export interface IpostDeletionModal {
+  isVisible: boolean;
+  isSubmit: boolean;
+  onPressRight: () => void;
+  onPressLeft: () => void;
+}
